third_party/WebKit/LayoutTests/bluetooth/server/connect/connection-succeeds.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
const test_desc = 'Device will connect';
let device;
promise_test(() => getConnectedHealthThermometerDevice()
  .then(_ => ({device} = _))
  .then(gatt => assert_true(device.gatt.connected))
  .then(() => navigator.bluetooth.test.allResponsesConsumed())
  .then(consumed => assert_true(consumed)), test_desc);
</script>
